requirements.txt
python-docx==1.1.0
markdown==3.5.2
pathlib2==2.3.7
colorama==0.4.6
